Fix filter, refactor menuListener
