+++ app/site/projects.liquid
@@ -226,10 +226,10 @@ layout: base
           {% if currentPage < totalPages | minus: 1 %}
             <li class="usa-pagination__item usa-pagination__arrow">
               <a href="/projects/page/{{ currentPage | plus: 1 }}/" class="usa-pagination__link usa-pagination__next-page" aria-label="Next page">
-                <span class="usa-pagination__link-text">Next </span>
                 <svg class="usa-icon" aria-hidden="true" role="img">
                   <use xlink:href="/assets/img/sprite.svg#navigate_next"></use>
                 </svg>
+                <span class="usa-pagination__link-text">Next</span>
               </a>
             </li>
           {% endif %}
