+++ app/src/js/modules/utilities.js
@@ -1,4 +1,4 @@
-export function addGlobalEventListener(type, selector, callback, parent = document) {
+ export function addGlobalEventListener(type, selector, callback, parent = document) {
   parent.addEventListener(type, e => {
     if (e.target.matches(selector)) {
       callback(e);
